Add I²C slave support
This reuses the base protocol code from the bootloader to handle the
framing and respond to commands, but does not define any commands yet
(it will always respond with COMMAND_NOT_SUPPORTED). It does support
general call reset, which allows the master to reset a slave to force it
into the bootloader again.