Fix hopper sensor analogReads
Turns out ATTinyCore does not support passing pin numbers to analogRead,
since the pin numbers and analog channel numbers are not equal, but do
overlap, which prevents analogRead from distinguishing them.

This changes the code to use the channel number. Unfortunately there is
no macro to translate from pin number to channel number, but the reverse
is possible, so the mapping is verified using a static_assert.